+++ glean-core/ios/GleanTests/GleanTests.swift
@@ -227,4 +227,60 @@ class GleanTests: XCTestCase {
             XCTAssertNil(error, "Test timed out waiting for upload: \(error!)")
         }
     }
+
+    func testSendingOfStartupBaselinePingWithAppLifetimeMetric() {
+        // Set the dirty flag.
+        glean_set_dirty_flag(true.toByte())
+
+        let stringMetric = StringMetricType(
+            category: "telemetry",
+            name: "app_lifetime",
+            sendInPings: ["baseline"],
+            lifetime: .application,
+            disabled: false
+        )
+        stringMetric.set("HELLOOOOO!")
+
+        // Set up the test stub based on the default telemetry endpoint
+        let host = URL(string: Configuration.Constants.defaultTelemetryEndpoint)!.host!
+        stub(condition: isHost(host)) { data in
+            let path = (data as NSURLRequest).url!
+            let parts = path.absoluteString.split(separator: "/")
+            XCTAssertEqual("baseline", parts[4])
+
+            let body = (data as NSURLRequest).ohhttpStubs_HTTPBody()
+            let json = try! JSONSerialization.jsonObject(with: body!, options: []) as? [String: Any]
+            XCTAssert(json != nil)
+
+            // Check for the "dirty_startup" flag
+            let pingInfo = json?["ping_info"] as? [String: Any]
+            XCTAssertEqual("dirty_startup", pingInfo?["reason"] as? String)
+
+            // Ensure there is only the expected locale string metric
+            let metrics = json?["metrics"] as? [String: Any]
+            let strings = metrics?["string"] as? [String: Any]
+            let metric = strings?["telemetry.app_lifetime"] as? String
+            XCTAssertEqual("HELLOOOOO!", metric)
+
+            DispatchQueue.main.async {
+                // let the response get processed before we mark the expectation fulfilled
+                self.expectation?.fulfill()
+            }
+
+            return OHHTTPStubsResponse(
+                jsonObject: [],
+                statusCode: 200,
+                headers: ["Content-Type": "application/json"]
+            )
+        }
+
+        expectation = expectation(description: "baseline ping received")
+
+        // Restart glean and don't clear the stores.
+        // This should trigger a baseline ping with a "dirty_startup" reason.
+        Glean.shared.resetGlean(clearStores: false)
+        waitForExpectations(timeout: 5.0) { error in
+            XCTAssertNil(error, "Test timed out waiting for upload: \(error!)")
+        }
+    }
 }