+++ glean-core/ios/GleanTests/Net/DeletionRequestPingTests.swift
@@ -67,7 +67,7 @@ class DeletionRequestPingTests: XCTestCase {
         glean.enableTestingMode()
 
         // Create directory for pending deletion-request pings
-        let pendingDeletionRequestDir = getDocumentsDirectory().appendingPathComponent("deletion_request")
+        let pendingDeletionRequestDir = getGleanDirectory().appendingPathComponent("deletion_request")
         try! FileManager.default.createDirectory(
             atPath: pendingDeletionRequestDir.path,
             withIntermediateDirectories: true,