+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v27.0.0...master)
+[Full changelog](https://github.com/mozilla/glean/compare/v27.1.0...master)
+
+# v27.1.0 (2020-04-09)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v27.0.0...v27.1.0)
+
+* General:
+  * BUGFIX: baseline pings sent at startup with the `dirty_startup` reason will now include application lifetime metrics ([#810](https://github.com/mozilla/glean/pull/810))
+* iOS:
+  * **Breaking change:** Change Glean iOS to use Application Support directory [#815](https://github.com/mozilla/glean/pull/815). No migration code is included. This will reset collected data if integrated without migration. Please [contact the Glean SDK team](https://github.com/mozilla/glean#contact) if this affects you.
+* Python
+  * BUGFIX: Fixed a race condition between uploading pings and deleting the temporary directory on shutdown of the process.
 
 # v27.0.0 (2020-04-08)